+++ src/Math/MadgwickAHRS.h
@@ -28,7 +28,7 @@ extern "C" {
 
 	extern void MadgwickAHRSupdate(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz);
 	extern void MadgwickAHRSupdateIMU(float gx, float gy, float gz, float ax, float ay, float az);
-	extern void QuaternionToEuler(double* roll, double* pitch, double* yaw);
+	extern void QuaternionToEuler(float* roll, float* pitch, float* yaw);
 
 #if __cplusplus
 }
